+++ test/service.test.js
@@ -1,10 +1,10 @@
 /* global describe, it, expect */
 
 var expect = require('chai').expect;
-var factory = require('../../../com/login/http/service');
+var factory = require('../com/service');
 
 
-describe('http/service', function() {
+describe('service', function() {
   
   it('should be annotated', function() {
     expect(factory['@implements']).to.deep.equal('http://i.bixbyjs.org/http/Service');